Add ToggleSection and LogoSection
