+++ apps/website/app/ui/sections/index.ts
@@ -7,3 +7,6 @@ export * from "./ModalSection";
 export * from "./TooltipSection";
 export * from "./TabsSection";
 export * from "./PopoverSection";
+export * from "./LogoSection";
+export * from "./ModalSection";
+export * from "./ToogleGroupSection";
